caibx: be more lenient about feature flag combinations
Older casync versions generate other feature flags in the caibx header.
We don't really care about that - the only one that currently is
relevant for us is whether the caibx file uses SHA512 or SHA256 hashing.

Instead of matching on an exact combination of feature flags, just
bail out on unknown/unexpected or prohibited ones.