store: always format chunk id
If a chunk is not found in any store, we print a warning which contains
the chunk's ID. However, this buffer is currently only filled when the
loglevel is "debug". This was an optimization that was valid because we
only used the stringified ID in debug messages originally. This causes
us to print uninitialized memory when hitting said warning.

Format the chunk ID unconditionally.