store-http: validate received chunks
When downloading chunks via HTTP, we currently trust that the contents
of a file match the hash advertised in its name. However, we want to be
able to give the guarantee that the resultant state of the target is
exactly equal to what the caibx file described. While bitrot could be
caught through zstd, this approach is not sound under adversarial
conditions. Thus, we need to rehash each downloaded chunk and check if
we got what we asked for.