store-http: also retry on resolver failures
Currently, we instantly abort when libcurl returns
CURLE_COULDNT_RESOLVE_HOST, but this could not only be caused by
NXDOMAIN or simliar (which would be relatively permanent and not
amenable to retries), but also by temporary network issues that cause
DNS timeouts. As we cannot distinguish between the two based on the
libcurl error code, be conservative and also go through the retry loop
in this case.