target: only build index when used as a store
We implicitly build a store out our target while creating it. The idea
is that this allows us to reuse chunks that occur multiple times without
having to re-fetch them from other (possibly remote) stores.

However, the benefits of caching using the target are small in practice
(the usual hit rate is about 1%) compared to the performance impact of
maintaining a live index. Only do this when we are in fact being used as
a store.