only use target as cache when explicitly requested
In almost all cases, using the target as a chunk cache is not useful
(syncing without a local store being a possible exception), but always
incurs a bookkeeping overhead.

Instead of doing this by default, require the user to explicitly request
this by specifying the target as a store on the command line.