skip chunk if already in place
While iterating through the chunks first check if the target already
contains the correct data at the chunk position before fetching the
chunk from any of the stores.

In case the buffer already contains the correct chunk data before the
check, we use a different buffer to avoid overwriting the data. Since
chunks are of variable size, it is not worth remembering the hash of
the data in the second buffer. It most likely does not end on a chunk
boundary after len bytes if it doesn't match the correct id.