skip matching prefix in target
When a synchronization process is interrupted for some reason and later
restarted, we will ignore the data that already exists and start from
the beginning again. This means that in the worst case, we will not make
any forward progress at all if the synchronization is interrupted again.

Instead, check the contents of the target to detect if matching data is
already present before fetching it from some other store.

Note that we only do this initially: Once a mismatch is detected, the
synchronization process proceeds as it did previously. This is for
performance reasons: While the case of resuming an interrupted
synchronization can be assumed to be relatively common, finding chunks
that are perfectly aligned at a later point in the target after a
difference has been detected earlier and *not* finding them in a local
chunk source is rather unlikely. The primary benefit to checking for
this condition all the time would be reducing redundant writes to the
target somewhat.

However, while the benefits to doing that are minor, the cost is rather
large: Now we are additionally hashing the whole target device.

Benchmarking this on the lowish-spec reference platform I use (Cortex-A5
SAMA5D2 attached to an eMMC) with a local chunk store only to avoid
network interference, this approach causes a slowdown of about 50% when
there are no matches on the target. If the target matches fully however,
it's about 30% faster.

Adding an early abort mechanism as described above lets us get the best
of both worlds: There is no measurable slowdown when the target doesn't
match and we get the same 30% speedup if it does, while also always
making forward progress when resuming an interrupted synchronization.

Thanks to Daniel Glöckner for the initial patch and discussion.