derive target index size from input
When converting a target to a store we currently use the same logic as
with creating a local store and base the size of the index on its size
in bytes. However, this does not make sense since by definition, the
target will accumulate the exact number of distinct chunks that are
present in the input caibx file. Furthermore, when writing to a regular
file instead of a block device (which is useful for regenerating a disk
image from a caibx file), the file will initially be empty, yielding an
estimate of 0, which is not permitted by index_init.

Fix both issues by simply taking the number of chunks in the input file
as an estimate. This will overestimate the size of the index since there
will usually be some redundancy in the input, but it's still going to be
a better estimate than what the current approach yields.