cmake: accept versions up to 3.5
CMake 3.27 deprecates compatibility with versions older than 3.5. While
versions >=3.5 are present in the currently-supported versions of
Buildroot and Yocto, let's not needlessly break setups that are stuck on
ancient versions for whatever reason.

CMake 3.12 introduced version ranges for the cmake_minimum_required()
command. This allows the CMakeLists.txt file to signal that it does not
require compatibility with versions older than <max>.

Since our CMakeLists.txt seems to work fine when targetting 3.5 (the
only difference that stands out is the -rdynamic change in 3.4/CMP0065,
but this is not affect us meaningfully), we can just leave <min> as is
and set <max> to 3.5.

Note however that since <max> was only introduced in 3.12 and older
versions ignore it, versions between 3.1 and 3.12 will still use
3.1 compatibility and thus we have to keep the CMP0063-statement in
place to keep versions in the 3.3 to 3.12 (exclusive) range happy.