store-http: support encryption
Add support for decrypting encrypted chunks stored on an HTTP store.

We do this by utilizing the elastic decompression buffer we use anyway
for the case where libzstd only partially processes the fragment of data
that came in from HTTP. When encryption is enabled, we simply pipe all
data through this buffer and decrypt the contents on the fly.

For supplying the key material, a special URL fragment parameter
(encrypt=) is added that passes its value to encrypt_parse_keyspec()
which allows loading the actual key from a number of sources.