allow building casync-nano with sanitizers
Add a CMake flag to build the main casync-nano tools with ASan/UBSan
enabled.

This is not really meant for production use, but rather to allow running
end-to-end tests with sanitizers. This is useful because the end-to-end
tests will by their nature exercise the code paths we expect to see in
real-world usage too.