target: support non-seekable targets
We currently require that whatever target we are writing to is seekable
and readable. These capabilities are necessary for a number of
scenarios:

 - Being able to write sparse images
 - Being able to use the target as a store as well
 - Resuming an interrupted synchronization process

Of those however, only the third is common. The first does not happen
with regular caibx files and the second is only useful in very specific
scenarios and carries a performance penalty.

Meanwhile, being able to write the image to a pipe or another kind of
file that isn't seekable or readable (such as UBI volumes once the
UBI_IOCVOLUP ioctl has been called, like RAUC does for example) is also
useful.

Add support for such targets by skipping the resumption logic and making
sure all writes are exactly consecutive.