fix includes for off_t
We are using the off_t type in various places and this is nominally
defined in sys/types.h, which we don't always include. For glibc/uClibc
this still worked because other headers pulled in the right definitions
transitively, but musl builds fail.

Just include the correct header whenerever off_t is used.